# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smartbiz Payroll API",
        "description": "API documentation for Smartbiz Payroll",
        "contact": {
            "email": "adisayoga@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://smartbiz-payroll.test/api/v1",
            "description": "Smartbiz Payroll"
        }
    ],
    "paths": {
...
        "/payroll-balances": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get all payroll balance per page",
                "operationId": "30cc7dcf7f36157696a64339be614659",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per-page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "{field: string, operator?: string, value: any}[]"
                        }
                    },
                    {
                        "name": "sorts",
                        "in": "query",
                        "schema": {
                            "type": "string[]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll balance",
                "operationId": "7bc654b6ff148586a57856eb20ddbd74",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-balances/{balanceId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll balance",
                "operationId": "d78fb4119857dfdcc06ec8c883877dbb",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll balance",
                "operationId": "66fa50f370171121b56729abe17ad071",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll balance",
                "operationId": "5ff8be8ba400eeb304386d0de1892461",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/values/{feedId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll balance feed",
                "operationId": "e4f5df860b909c8fe6bb213dfe6e3d51",
                "parameters": [
                    {
                        "name": "feedId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-balances/{balanceId}/feeds": {
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll balance feed",
                "operationId": "6fb67e8e1b482d5633d7f71d980e405d",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-balances/feeds/{balanceId}": {
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll balance feed",
                "operationId": "cee0b008daf7e735fd0b9dfb8b29b1f3",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll balance feed",
                "operationId": "042e0b877bad5ba60f16acc61577074e",
                "parameters": [
                    {
                        "name": "balanceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-elements/values/{inputValueId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll input value",
                "operationId": "e9edbec59f98051bfcdc124c76f26a6a",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll input value",
                "operationId": "a45e3493738026f793f56fb34ff9bbf6",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll input value",
                "operationId": "a152d2c9c8787cca08070fb4275b2c2d",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-elements/values/{resultId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll formula result",
                "operationId": "9733cbdf3d48aa563f6a1dcb876e4d5d",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer [token]",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}